+++ package.json
@@ -7,6 +7,7 @@
     "lintTS": "tsc --noemit && eslint . --ext .ts,.tsx",
     "start": "cross-env NODE_ENV=development webpack serve --config webpack.config.js",
     "build": "cross-env NODE_ENV=development webpack --config webpack.config.js",
+    "serve": "npx http-server -c-1 ./dist",
     "lint": "eslint 'src/**/*.js' || true",
     "prettify": "prettier --write 'src/**/*.js'"
   },
@@ -37,7 +38,7 @@
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^9.0.1",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.4.0",
+    "css-loader": "^6.8.1",
     "css-minimizer-webpack-plugin": "^3.1.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-typescript": "^12.3.1",
@@ -56,13 +57,13 @@
     "prettier": "^2.4.1",
     "redux-loop": "6.1.0",
     "sass": "^1.43.5",
-    "sass-loader": "^12.2.0",
-    "style-loader": "^3.3.0",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
     "ts-loader": "^9.3.1",
     "typedoc": "^0.17.8",
     "typedoc-plugin-markdown": "^2.4.2",
     "typescript": "^4.3.5",
-    "webpack": "^5.58.2",
+    "webpack": "^5.89.0",
     "webpack-cli": "^4.9.0",
     "webpack-dev-server": "^4.11.1",
     "webpack-merge": "^5.8.0"
