+++ src/components/BloombergToFDC3/common.ts
