+++ src/components/BloombergToFDC3/components/SecuritySearch.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { FormEvent, useState } from "react";
+import React, { FormEvent, useState } from "react";
 import Autosuggest, {
 	FetchRequestedReasons,
 	InputProps,
