--- src/components/BloombergToFDC3/components/SecuritySearchBox.tsx
@@ -1,14 +1,0 @@
-import React, {ChangeEventHandler} from "react";
-
-type SecuritySearchBoxProps = {
-	value: string;
-	changeFunction: ChangeEventHandler;
-}
-
-export const SecuritySearchBox = ({
-							   value,
-							   changeFunction
-						   }:SecuritySearchBoxProps) => {
-
-	return <div><input type="text" value={value} onChange={changeFunction} placeholder="Security search box" /></div>;
-}
