+++ src/components/BloombergToFDC3/config.json
@@ -1,60 +1,25 @@
 {
-	appId: "Bloomberg Security Finder",
-	name: "Bloomberg Security Finder",
-	description: "An interface between Finsemble windows and Bloomberg, passing Securities through and doing lookups.",
-	manifest: {
-		window: {
-			url: "$applicationRoot/components/SecurityFinder/index.html",
-			addToWorkspace: true,
-			options: {
-				resizable: true,
-				autoShow: true,
-				alwaysOnTop: false,
-				securityPolicy: "trusted",
-            },
-			top: "center",
-			left: "center",
-			width: 450,
-			height: 400,
-        },
-		component: {
-			preload: [
-                "$servicesRoot/bloombergBridge/BloombergBridgePreload.js"
-            ],
-			spawnOnStartup: false,
-			singleton: false,
-			showDevConsoleOnVisible: false,
-        },
-		foreign: {
-			services: {
-				windowService: {
-					manageWindowMovement: true,
-					allowAutoArrange: false,
-					allowSnapping: true,
-					allowTabbing: true,
-					allowTiling: true,
-					allowGrouping: true,
-					allowMinimize: true,
-                },
-				workspace: {},
-            },
-			components: {
-                "App Launcher": {
-					launchableByUser: true,
-                },
-                "Window Manager": {
-					alwaysOnTopIcon: true,
-					title: "BBG Security Finder",
-					persistWindowState: true,
-					hideMaximize: true,
-					showLinker: true,
-                },
-				Toolbar: {
-					iconClass: "ff-search",
-                },
-            },
-        },
-    },
-	version: "1.0.0",
-	tooltip: "Bloomberg Security Finder",
+	"appId": "BloombergFDC3Demo",
+	"name": "Bloomberg FDC3 Demo",
+	"type": "web",
+	"details": {
+		"url": "http://localhost:5173/"
+	},
+	"hostManifests": {
+		"Finsemble": {
+			"component": {
+				"preload": ["$moduleRoot/preloads/BloombergBridgePreload.js"]
+			}
+		},
+		"foreign": {
+			"components": {
+				"App Launcher": {
+					"launchableByUser": true
+				},
+				"Window Manager": {
+					"titlebarType": "injected"
+				}
+			}
+		}
+	}
 }
