+++ src/components/BloombergToFDC3/index.html
@@ -5,7 +5,7 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<meta name="description" content="Desktop search widget for Bloomberg Securities" />
 		<title>Bloomberg Security Finder</title>
-		<link rel="stylesheet" type="text/css" href="css/theme.css" />
+		<link rel="stylesheet" type="text/css" href="css/App.css" />
 	</head>
 	<body>
 		<div id="root">Waiting for Bloomberg...</div>
