+++ tsconfig.json
@@ -4,16 +4,19 @@
 		"allowSyntheticDefaultImports": true,
 		"jsx": "react",
 		"moduleResolution": "node",
-		"noImplicitAny": true,
 		"preserveConstEnums": true,
 		"removeComments": true,
 		"sourceMap": true,
-		"strict": true,
 		"target": "ES2020",
 		"typeRoots": [
 			"node_modules/@types",
 			"node_modules/@finsemble"
-		]
+		],
+		/* Linting */
+		"strict": true,
+		"noUnusedLocals": false,
+		"noUnusedParameters": false,
+		"noImplicitAny": false
 	},
 	"include": [
 		"src/components/Bloomberg Demo/*"
